feat(discover): 将推荐歌单替换为最近播放列表
- 替换推荐歌单功能为最近播放歌曲列表
- 修改API域名配置默认值从192.168.0.13改为192.168.0.23
- 在API域名对话框中显示当前域名而非仅作为提示
- 添加歌曲列表页面和相关路由配置
- 创建歌曲列表项绑定器用于显示最近播放的歌曲
- 修改发现页面UI组件ID从recommendPlaylist改为songHistory
- 更新播放计数字段类型从Int改为Long以支持更大的数值
- 添加歌单历史列表缓存机制
- 实现不同展示类型的歌曲列表（全部和历史）

+++ app/src/main/java/me/wcy/music/discover/playlist/detail/bean/NmSongData.kt
@@ -57,7 +57,7 @@ data class NmSongData(
     @SerializedName("path")
     var path: String = "",
     @SerializedName("playCount")
-    var playCount: Int = 0,
+    var playCount: Long = 0,
     @SerializedName("playDate")
     var playDate: Any? = Any(),
     @SerializedName("rating")
